+++ jest.config.js
@@ -78,10 +78,9 @@ module.exports = {
   ],
 
   // A map from regular expressions to module names that allow to stub out resources with a single module
-  // moduleNameMapper: {
-  //   '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': '<rootDir>/__mocks__/fileMock.js',
-  //   '\\.(css|less)$': '<rootDir>/__mocks__/styleMock.js'
-  // },
+  moduleNameMapper: {
+    '^shared(.*)$': '/Users/talekperov/latest_homs/homs/hbw/app/javascript/packs/hbw/components/shared$1'
+  },
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
   // modulePathIgnorePatterns: [],